+++ compiler/hash-ast/Cargo.toml
@@ -18,3 +18,5 @@ strum_macros = "0.21"
 derive_more = "0.99"
 hash-utils = { path = "../hash-utils" } 
 hash-alloc = { path = "../hash-alloc" }
+hash-source = { path = "../hash-source" }
+hash-reporting = {path = "../hash-reporting" }
