+++ compiler/hash-parser/src/lexer.rs
@@ -4,10 +4,10 @@
 //! All rights reserved 2021 (c) The Hash Language authors
 use hash_alloc::{collections::row::Row, row, Wall};
 use hash_ast::error::ParseResult;
+use hash_ast::ident::IDENTIFIER_MAP;
 use hash_ast::literal::STRING_LITERAL_MAP;
-use hash_ast::location::Location;
 use hash_ast::{ident::CORE_IDENTIFIERS, keyword::Keyword};
-use hash_ast::{ident::IDENTIFIER_MAP, module::ModuleIdx};
+use hash_source::{location::Location, module::ModuleIdx};
 
 use crate::{
     error::{TokenError, TokenErrorKind, TokenErrorWrapper},
