+++ compiler/hash-reporting/src/errors.rs
@@ -2,16 +2,11 @@
 //!
 //! All rights reserved 2021 (c) The Hash Language authors
 
-use hash_ast::{error::ParseError, ident::IDENTIFIER_MAP};
-use hash_typecheck::{error::TypecheckError, storage::GlobalStorage, writer::TypeWithStorage};
 use std::fmt;
 use std::{io, process::exit};
 use thiserror::Error;
 
-use crate::{
-    highlight::{highlight, Colour, Modifier},
-    reporting::{Report, ReportBuilder, ReportCodeBlock, ReportElement, ReportKind, ReportNote},
-};
+use crate::highlight::{highlight, Colour, Modifier};
 
 /// Enum representing the variants of error that can occur when running an interactive session
 #[derive(Error, Debug)]
@@ -44,312 +39,6 @@ impl fmt::Display for ErrorCode {
         write!(f, "{:0>4}", *self as u32)
     }
 }
-
-impl From<(TypecheckError, GlobalStorage<'_, '_, '_>)> for Report {
-    fn from((error, storage): (TypecheckError, GlobalStorage<'_, '_, '_>)) -> Self {
-        let mut builder = ReportBuilder::new();
-        builder
-            .with_kind(ReportKind::Error)
-            .with_message("Failed to typecheck")
-            .with_error_code(ErrorCode::Typecheck); // @@ErrorReporting: Get the correct typecheck code
-
-        match error {
-            TypecheckError::TypeMismatch(left, right) => {
-                let left_ty = TypeWithStorage::new(left, &storage);
-                let right_ty = TypeWithStorage::new(right, &storage);
-
-                // @@TODO: we want to print the location of the lhs_expression where the type mismatches
-                //         and the right hand side
-                builder.add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    format!(
-                        "Types mismatch, got a `{}`, but wanted a `{}`.",
-                        left_ty, right_ty
-                    ),
-                )));
-            }
-            TypecheckError::UsingBreakOutsideLoop(src) => {
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(src, "here")))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        "You can't use a `break` clause outside of a loop.",
-                    )));
-            }
-            TypecheckError::UsingContinueOutsideLoop(src) => {
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(src, "here")))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        "You can't use a `continue` clause outside of a loop.",
-                    )));
-            }
-            TypecheckError::UsingReturnOutsideFunction(src) => {
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(src, "here")))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        "You can't use a `return` clause outside of a function body.",
-                    )));
-            }
-            TypecheckError::RequiresIrrefutablePattern(src) => {
-                builder
-                .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(src, "This pattern isn't refutable")))
-                .add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    "Destructuring statements in `let` or `for` statements must use an irrefutable pattern.",
-                )));
-            }
-            TypecheckError::UnresolvedSymbol(symbol) => {
-                let ident_path = symbol.get_ident();
-                let formatted_symbol = format!("{}", IDENTIFIER_MAP.get_path(ident_path));
-
-                if let Some(location) = symbol.location() {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "Unresolved symbol",
-                    )));
-                }
-
-                // At-least we can print the symbol that wasn't found...
-                builder.add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    format!(
-                        "Symbol `{}` is not defined in the current scope.",
-                        formatted_symbol
-                    ),
-                )));
-            }
-            TypecheckError::TryingToNamespaceType(symbol) => {
-                let symbol_name = IDENTIFIER_MAP.get_path(symbol.get_ident());
-
-                if let Some(location) = symbol.location() {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        format!(
-                            "This symbol `{}` is defined as a type in the current scope.",
-                            symbol_name
-                        ),
-                    )));
-                }
-
-                builder.add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    "You cannot namespace a symbol that's a type.",
-                )));
-            }
-            TypecheckError::TryingToNamespaceVariable(symbol) => {
-                let symbol_name = IDENTIFIER_MAP.get_path(symbol.get_ident());
-
-                if let Some(location) = symbol.location() {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "This is a variable",
-                    )));
-                }
-
-                builder.add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    format!("`{}` is a variable. You cannot namespace a variable defined in the current scope.", symbol_name),
-                )));
-            }
-            TypecheckError::UsingVariableInTypePos(symbol) => {
-                let symbol_name = IDENTIFIER_MAP.get_path(symbol.get_ident());
-
-                if let Some(location) = symbol.location() {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "This is expects a type instead of a variable.",
-                    )));
-                }
-
-                builder.add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    format!("`{}` is a variable and not a type. You cannot use a variable in the place of a type.", symbol_name),
-                )));
-            }
-            TypecheckError::UsingTypeInVariablePos(symbol) => {
-                let symbol_name = IDENTIFIER_MAP.get_path(symbol.get_ident());
-                
-                if let Some(location) = symbol.location() {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "You can't use a type here...",
-                    )));
-                }
-
-                builder.add_element(ReportElement::Note(ReportNote::new(
-                    "note",
-                    format!("`{}` is a type and not a variable. You cannot use a type in the place of a variable.", symbol_name),
-                )));
-            }
-            TypecheckError::TypeIsNotStruct {
-                ty,
-                location,
-                ty_def_location,
-            } => {
-                let ty = TypeWithStorage::new(ty, &storage);
-
-                // Print where the original type is defined with an annotation.
-                if let Some(ty_def_location) = ty_def_location {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        ty_def_location,
-                        format!("The type `{}` is defined here.", ty),
-                    )));
-                }
-
-                // Print where the literal being created...
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "Not a struct",
-                    )))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        format!("This type `{}` isn't a struct.", ty),
-                    )));
-            }
-            TypecheckError::UnresolvedStructField {
-                ty_def_location,
-                ty_def_name,
-                field_name,
-                location,
-            } => {
-                let name = IDENTIFIER_MAP.get_ident(field_name);
-                let ty_name = IDENTIFIER_MAP.get_ident(ty_def_name);
-
-                // If we have the location of the definition, we can print it here
-                if let Some(ty_def_location) = ty_def_location {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        ty_def_location,
-                        format!("The struct `{}` is defined here.", ty_name),
-                    )));
-                }
-
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "Unknown field",
-                    )))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        format!("The field `{}` doesn't exist on struct `{}`.", name, name),
-                    )));
-            }
-            TypecheckError::ExpectingBooleanInCondition { found, location } => {
-                let found_ty = TypeWithStorage::new(found, &storage);
-
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "Expression should be of `boolean` type",
-                    )))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        format!("In `if` statements, the condition must be explicitly of `boolean` type, however the expression was found to be of `{}` type.", found_ty),
-                    )));
-            }
-            TypecheckError::MissingStructField {
-                ty_def_location,
-                ty_def_name,
-                field_name,
-                field_location,
-            } => {
-                let name = IDENTIFIER_MAP.get_ident(field_name);
-                let ty_name = IDENTIFIER_MAP.get_ident(ty_def_name);
-
-                // If we have the location of the definition, we can print it here
-                if let Some(ty_def_location) = ty_def_location {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        ty_def_location,
-                        format!("The struct `{}` is defined here.", ty_name),
-                    )));
-                }
-
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        field_location,
-                        "Struct is missing field",
-                    )))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        format!("The struct `{}` is missing the field `{}`.", ty_name, name),
-                    )));
-            }
-            TypecheckError::BoundRequiresStrictlyTypeVars(location) => {
-                // @@TODO: Maybe report here what we found?
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location, "here",
-                    )))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        "This type bound should only contain type variables",
-                    )));
-            }
-            TypecheckError::InvalidPropertyAccess {
-                field_name,
-                location,
-                ty_def_name,
-                ty_def_location,
-            } => {
-                let name = IDENTIFIER_MAP.get_ident(field_name);
-                let ty_name = IDENTIFIER_MAP.get_ident(ty_def_name);
-
-                // If we have the location of the definition, we can print it here
-                if let Some(ty_def_location) = ty_def_location {
-                    builder.add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        ty_def_location,
-                        format!("The struct `{}` is defined here.", ty_name),
-                    )));
-                }
-
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(
-                        location,
-                        "unknown field",
-                    )))
-                    .add_element(ReportElement::Note(ReportNote::new(
-                        "note",
-                        format!("The field `{}` doesn't exist on type `{}`.", name, ty_name),
-                    )));
-            }
-        }
-
-        // @@ErrorReporting: we might want to properly handle incomplete reports?
-        builder.build().unwrap()
-    }
-}
-
-impl From<ParseError> for Report {
-    fn from(error: ParseError) -> Self {
-        let mut builder = ReportBuilder::new();
-        builder
-            .with_kind(ReportKind::Error)
-            .with_message("Failed to parse")
-            .with_error_code(ErrorCode::Parsing);
-
-        match error {
-            ParseError::Parsing {
-                message,
-                src: Some(src),
-            }
-            | ParseError::Token { message, src } => {
-                builder
-                    .add_element(ReportElement::CodeBlock(ReportCodeBlock::new(src, "here")))
-                    .add_element(ReportElement::Note(ReportNote::new("note", message)));
-            }
-            // When we don't have a source for the error, just add a note
-            ParseError::Parsing { message, src: None } => {
-                builder.with_message(message);
-            }
-        };
-
-        // @@ErrorReporting: we might want to properly handle incomplete reports?
-        builder.build().unwrap()
-    }
-}
-
 /// Errors that might occur when attempting to compile and or interpret a
 /// program.
 #[derive(Debug, Error)]
