+++ compiler/hash-source/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "hash-reporting"
+name = "hash-source"
 version = "0.1.0"
 authors = ["The Hash Language authors"]
 edition = "2021"
 
 [dependencies]
 thiserror = "1.0"
-hash-source = {path = "../hash-source" }
-hash-alloc = {path = "../hash-alloc" }
+derive_more = "0.99"
+hash-utils = { path = "../hash-utils" }
