+++ compiler/hash-source/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod location;
+pub mod module;
