+++ compiler/hash-typecheck/Cargo.toml
@@ -6,8 +6,10 @@ edition = "2021"
 
 [dependencies]
 hash-ast = { path = "../hash-ast" }
+hash-source = { path = "../hash-source" }
 hash-utils = { path = "../hash-utils" }
 hash-alloc = { path = "../hash-alloc" }
+hash-reporting = {path = "../hash-reporting" }
 smallvec = { version = "1.6", features = ["union"] }
 slotmap = "1.0"
 dashmap = "4.0"
